backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.0",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=4.3.1",
    "python-dotenv>=1.0.0",
]
